+++ src/main/java/com/timeToast/timeToast/controller/payment/PaymentAdminController.java
@@ -1,7 +1,7 @@
 package com.timeToast.timeToast.controller.payment;
 
-import com.timeToast.timeToast.dto.icon.icon_group.response.admin.IconGroupMonthlyRevenues;
-import com.timeToast.timeToast.dto.icon.icon_group.response.admin.IconGroupSummaries;
+import com.timeToast.timeToast.dto.payment.response.IconGroupMonthlyRevenues;
+import com.timeToast.timeToast.dto.payment.response.IconGroupSummaries;
 import com.timeToast.timeToast.dto.payment.response.PaymentDetailResponse;
 import com.timeToast.timeToast.dto.payment.response.PaymentsAdminResponses;
 import com.timeToast.timeToast.service.payment.PaymentService;
