+++ src/main/java/com/timeToast/timeToast/controller/payment/PaymentController.java
@@ -1,7 +1,6 @@
 package com.timeToast.timeToast.controller.payment;
 
 import com.timeToast.timeToast.domain.member.member.LoginMember;
-import com.timeToast.timeToast.dto.icon.icon_group.response.admin.IconGroupSummaries;
 import com.timeToast.timeToast.dto.payment.request.PaymentSaveRequest;
 import com.timeToast.timeToast.dto.payment.request.PaymentSuccessRequest;
 import com.timeToast.timeToast.dto.payment.response.PaymentFailResponse;
