+++ src/main/java/com/timeToast/timeToast/dto/icon/icon_group/response/IconGroupDetailResponses.java
@@ -1,4 +1,4 @@
-package com.timeToast.timeToast.dto.icon.icon_group.response.member;
+package com.timeToast.timeToast.dto.icon.icon_group.response;
 
 import java.util.List;
