+++ src/main/java/com/timeToast/timeToast/dto/icon/icon_group/response/IconGroupInfoResponse.java
@@ -1,0 +1,7 @@
+package com.timeToast.timeToast.dto.icon.icon_group.response;
+
+public record IconGroupInfoResponse(
+        IconGroupSummaryInfo iconGroupSummaryInfo,
+        boolean isBuy
+) {
+}
