--- src/main/java/com/timeToast/timeToast/dto/icon/icon_group/response/member/IconGroupInfoResponse.java
@@ -1,9 +1,0 @@
-package com.timeToast.timeToast.dto.icon.icon_group.response.member;
-
-import com.timeToast.timeToast.dto.icon.icon_group.response.IconGroupSummaryInfo;
-
-public record IconGroupInfoResponse(
-        IconGroupSummaryInfo iconGroupSummaryInfo,
-        boolean isBuy
-) {
-}
