+++ src/main/java/com/timeToast/timeToast/service/payment/PaymentServiceImpl.java
@@ -11,10 +11,10 @@ import com.timeToast.timeToast.domain.icon.icon_member.IconMember;
 import com.timeToast.timeToast.domain.member.member.Member;
 import com.timeToast.timeToast.domain.payment.Payment;
 import com.timeToast.timeToast.domain.premium.Premium;
-import com.timeToast.timeToast.dto.icon.icon_group.response.admin.IconGroupMonthlyRevenue;
-import com.timeToast.timeToast.dto.icon.icon_group.response.admin.IconGroupMonthlyRevenues;
-import com.timeToast.timeToast.dto.icon.icon_group.response.admin.IconGroupSummaries;
-import com.timeToast.timeToast.dto.icon.icon_group.response.admin.IconGroupSummary;
+import com.timeToast.timeToast.dto.payment.response.IconGroupMonthlyRevenue;
+import com.timeToast.timeToast.dto.payment.response.IconGroupMonthlyRevenues;
+import com.timeToast.timeToast.dto.payment.response.IconGroupSummaries;
+import com.timeToast.timeToast.dto.payment.response.IconGroupSummary;
 import com.timeToast.timeToast.dto.payment.IconGroupPaymentSummaryDto;
 import com.timeToast.timeToast.dto.payment.request.PaymentSaveRequest;
 import com.timeToast.timeToast.dto.payment.request.PaymentSuccessRequest;
