+++ src/test/java/com/timeToast/timeToast/service/icon/icon_group/IconServiceImplTest.java
@@ -5,9 +5,9 @@ import com.timeToast.timeToast.domain.icon.icon_group.IconGroup;
 import com.timeToast.timeToast.domain.icon.icon_member.IconMember;
 import com.timeToast.timeToast.domain.member.member.Member;
 import com.timeToast.timeToast.dto.icon.icon.IconResponse;
-import com.timeToast.timeToast.dto.icon.icon_group.response.member.IconGroupDetail;
-import com.timeToast.timeToast.dto.icon.icon_group.response.member.IconGroupInfoResponses;
-import com.timeToast.timeToast.dto.icon.icon_group.response.member.IconGroupDetailResponses;
+import com.timeToast.timeToast.dto.icon.icon_group.response.IconGroupDetail;
+import com.timeToast.timeToast.dto.icon.icon_group.response.IconGroupInfoResponses;
+import com.timeToast.timeToast.dto.icon.icon_group.response.IconGroupDetailResponses;
 import com.timeToast.timeToast.global.constant.StatusCode;
 import com.timeToast.timeToast.global.exception.NotFoundException;
 import com.timeToast.timeToast.global.response.Response;
