+++ src/test/java/com/timeToast/timeToast/service/payment/PaymentServiceTest.java
@@ -3,10 +3,10 @@ package com.timeToast.timeToast.service.payment;
 import com.timeToast.timeToast.domain.enums.icon_group.IconType;
 import com.timeToast.timeToast.domain.enums.payment.ItemType;
 import com.timeToast.timeToast.domain.enums.payment.PaymentState;
-import com.timeToast.timeToast.dto.icon.icon_group.response.admin.IconGroupMonthlyRevenue;
-import com.timeToast.timeToast.dto.icon.icon_group.response.admin.IconGroupMonthlyRevenues;
-import com.timeToast.timeToast.dto.icon.icon_group.response.admin.IconGroupSummaries;
-import com.timeToast.timeToast.dto.icon.icon_group.response.admin.IconGroupSummary;
+import com.timeToast.timeToast.dto.payment.response.IconGroupMonthlyRevenue;
+import com.timeToast.timeToast.dto.payment.response.IconGroupMonthlyRevenues;
+import com.timeToast.timeToast.dto.payment.response.IconGroupSummaries;
+import com.timeToast.timeToast.dto.payment.response.IconGroupSummary;
 import com.timeToast.timeToast.dto.payment.request.PaymentSaveRequest;
 import com.timeToast.timeToast.dto.payment.request.PaymentSuccessRequest;
 import com.timeToast.timeToast.dto.payment.response.*;
